+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- `Source::map()`
+- `Source::flatten()`
+- `impl Default for Cell`
+- `impl Default for CellWithLocal`
+- `impl Listen for Cell`
+- `impl Listen for CellWithLocal`
+
 ## 0.1.0 (2024-12-19)
 
 Initial public release.